fix signedness warnings

+++ include/transpose.h
@@ -254,8 +254,8 @@ class Transpose
       int dim_; //!< dimension of the tensor
       std::vector<size_t> sizeA_; //!< size of A
       std::vector<int> perm_; //!< permutation 
-      std::vector<size_t> outerSizeA_; //!< outer sizes of A
-      std::vector<size_t> outerSizeB_;  //!< outer sizes of B
+      std::vector<int> outerSizeA_; //!< outer sizes of A
+      std::vector<int> outerSizeB_;  //!< outer sizes of B
       std::vector<size_t> lda_;  //!< strides for all dimensions of A (first dimension has a stride of 1)
       std::vector<size_t> ldb_;  //!< strides for all dimensions of B (first dimension has a stride of 1)
       std::vector<int> threadIds_; //!< OpenMP threadIds of the threads involed in the transposition
